% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{Counter}
\alias{Counter}
\title{Counter R6 class to represent a counter}
\description{
Counter R6 class to represent a counter

Counter R6 class to represent a counter
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{number of times that something has
happened}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Counter$new()}}
\item \href{#method-print}{\code{Counter$print()}}
\item \href{#method-increment}{\code{Counter$increment()}}
\item \href{#method-chain_increment}{\code{Counter$chain_increment()}}
\item \href{#method-pipe_increment}{\code{Counter$pipe_increment()}}
\item \href{#method-reset}{\code{Counter$reset()}}
\item \href{#method-clone}{\code{Counter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Counter object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Counter} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Counter object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{required for compatibility with print()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-increment"></a>}}
\if{latex}{\out{\hypertarget{method-increment}{}}}
\subsection{Method \code{increment()}}{
Add one to counter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$increment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chain_increment"></a>}}
\if{latex}{\out{\hypertarget{method-chain_increment}{}}}
\subsection{Method \code{chain_increment()}}{
Add one to counter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$chain_increment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pipe_increment"></a>}}
\if{latex}{\out{\hypertarget{method-pipe_increment}{}}}
\subsection{Method \code{pipe_increment()}}{
Add one to counter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$pipe_increment(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{piped dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{.data}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Reset counter to zero.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$reset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
