% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{History}
\alias{History}
\title{History R6 class to represent historical values}
\description{
History R6 class to represent historical values

History R6 class to represent historical values
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{values}}{list of stored values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{History$new()}}
\item \href{#method-print}{\code{History$print()}}
\item \href{#method-log}{\code{History$log()}}
\item \href{#method-chain_log}{\code{History$chain_log()}}
\item \href{#method-pipe_log}{\code{History$pipe_log()}}
\item \href{#method-reset}{\code{History$reset()}}
\item \href{#method-clone}{\code{History$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new History object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{History$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{History} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print summary of \code{Counter} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{History$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{unused}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
Store a value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{History$log(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{value to add to history}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chain_log"></a>}}
\if{latex}{\out{\hypertarget{method-chain_log}{}}}
\subsection{Method \code{chain_log()}}{
Store a value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{History$chain_log(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{value to add to history}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pipe_log"></a>}}
\if{latex}{\out{\hypertarget{method-pipe_log}{}}}
\subsection{Method \code{pipe_log()}}{
Store a value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{History$pipe_log(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{value to add to history}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{val}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Clear history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{History$reset(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{pipeline value passed in (Optional)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{History$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
