% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimeter.R
\name{mm_get_identity_meter}
\alias{mm_get_identity_meter}
\title{Simple meter that does not transform its input and
performs no comparison of results other than returning
them.}
\usage{
mm_get_identity_meter()
}
\value{
\code{Multimeter} instance
}
\description{
Simple meter that does not transform its input and
performs no comparison of results other than returning
them.
}
\examples{
if (interactive()) {
  m <- mm_get_identity_meter()
  m$probe(1L) == 1L
  m$probe(2L) == 2L
  m$comparison
  print(m)
}
}
