% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimeter.R
\name{mm_get_names_meter}
\alias{mm_get_names_meter}
\title{Track column names before and after a transformation}
\usage{
mm_get_names_meter()
}
\value{
\code{Multimeter} instance
}
\description{
Track column names before and after a transformation
}
\examples{
if (interactive()) {
  m <- mm_get_names_meter()
  library(dplyr)
  mtcars2 <- mtcars \%>\%
    m$probe() \%>\%
    mutate(square_error_from_five = (cyl - 5)^2) \%>\%
    m$probe()
  print(m$comparison)
}
}
