% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimeter.R
\name{mm_get_value_meter}
\alias{mm_get_value_meter}
\title{Track changes in columns provided to $probe() function}
\usage{
mm_get_value_meter()
}
\value{
\code{Multimeter} instance
}
\description{
Track changes in columns provided to $probe() function
}
\examples{
if (interactive()) {
  m <- mm_get_state_capture_meter()
  library(dplyr)
  mtcars2 <- mtcars \%>\%
    m$probe(cyl) \%>\%
    mutate(square_error_from_five = (cyl - 5)^2) \%>\%
    m$probe(square_error_from_five)
  print(m$comparison)
}
}
