% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimeter-package.R
\docType{package}
\name{multimeter-package}
\alias{multimeter}
\alias{multimeter-package}
\title{multimeter: Inspect Data Pipelines}
\description{
Inspect data pipelines (especially magrittr pipelines) for before-after changes during tricky operations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bvancil/multimeter}
  \item Report bugs at \url{https://github.com/bvancil/multimeter/issues}
}

}
\author{
\strong{Maintainer}: Brian Vancil \email{1388508+bvancil@users.noreply.github.com} (\href{https://orcid.org/0000-0003-1145-3681}{ORCID})

}
\keyword{internal}
