% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimeter.R
\name{Multimeter}
\alias{Multimeter}
\title{Multimeter R6 class to represent a multimeter type}
\description{
Multimeter R6 class to represent a multimeter type

Multimeter R6 class to represent a multimeter type
}
\details{
A new Multimeter can insert probes in various stages
of a data pipeline, allowing you to compare output
before and after a change. To facilitate use with large
datasets for which you might only wish to store a summary
of the data, a Multimeter object can take a \code{transform}
function that transforms the data before storing it, and
a \code{compare} function to point out salient differences.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transform}}{function of dataset}

\item{\code{compare}}{function of before and after
transformed datasets}

\item{\code{before}}{transformed "before" dataset}

\item{\code{after}}{transformed "after" dataset}

\item{\code{comparison}}{result of \code{compare} on \code{before}
and \code{after}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Multimeter$new()}}
\item \href{#method-print}{\code{Multimeter$print()}}
\item \href{#method-probe}{\code{Multimeter$probe()}}
\item \href{#method-save_memory}{\code{Multimeter$save_memory()}}
\item \href{#method-reset}{\code{Multimeter$reset()}}
\item \href{#method-clone}{\code{Multimeter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Multimeter object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multimeter$new(transformer, comparator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transformer}}{function of dataset to store}

\item{\code{comparator}}{function of transformed dataset to
compare}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Multimeter} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print summary of \code{Multimeter} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multimeter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{unused}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-probe"></a>}}
\if{latex}{\out{\hypertarget{method-probe}{}}}
\subsection{Method \code{probe()}}{
Insert a probe at this stage of data pipeline.

Because this operation returns the first argument, it
may be used in magrittr or base R pipelines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multimeter$probe(.data, ..., .print_comparison = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{dataset tibble or data.frame}

\item{\code{...}}{other arguments passed on to
\code{self$transform}}

\item{\code{.print_comparison}}{logical whether to print a
comparison on the second \code{probe}. Ignored otherwise.
Default: \code{TRUE}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{.data}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_memory"></a>}}
\if{latex}{\out{\hypertarget{method-save_memory}{}}}
\subsection{Method \code{save_memory()}}{
Delete intermediate stored transformed before/after
datasets to save memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multimeter$save_memory(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{dataset Optional if used in a data pipeline}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{.data} (or \code{NULL} if \code{.data} is missing)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Reset multimeter for repeated use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multimeter$reset(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{dataset Optional if used in a data pipeline}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{.data} (or \code{NULL} if \code{.data} is missing)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multimeter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
